chore: move use-hover and use-longpress to hooks as well
