+++ packages/components/src/hooks/use-onclickoutside.js
@@ -1,14 +1,12 @@
 import React from 'react';
 
-function useOnClickOutside(ref, handler, validationFn) {
+export const useOnClickOutside = (ref, handler, validationFn) => {
     React.useEffect(() => {
         const listener = event => {
             const path = event.path ?? event.composedPath?.();
 
-            if (
-                ref &&
-                (!ref.current.contains(event.target) && !ref.current.contains(path && path[0])) // When component is isolated (e.g, iframe, shadow DOM) event.target refers to whole container not the component. path[0] is the node that the event originated from, it does not need to walk the array
-            ) {
+            // When component is isolated (e.g, iframe, shadow DOM) event.target refers to whole container not the component. path[0] is the node that the event originated from, it does not need to walk the array
+            if (ref && (!ref.current.contains(event.target) && !ref.current.contains(path && path[0]))) {
                 if (validationFn && !validationFn(event)) return;
                 handler(event);
             }
@@ -20,6 +18,4 @@ function useOnClickOutside(ref, handler, validationFn) {
             document.removeEventListener('mousedown', listener);
         };
     }, [ref, handler]);
-}
-
-export { useOnClickOutside };
+};
